test(desktop): isolate sidecar durability coverage
Problem:
The new session-sidecar regressions pushed an already large legacy test file further beyond the repository test-file-size budget.

Root cause:
The focused coverage was added beside older session tests instead of in its own owner-level test file.

Fix:
Move the new durability and concurrent-title tests unchanged into session_sidecar_durability_test.go.

Verification:
- go run ./tools/repolint
- cd desktop && go test -race . -run 'TestSessionSidecarSavesUseDurableAtomicWrite|TestSetSessionTitle|TestRecordSessionDisplaySerializesConcurrentTabs|TestRecordSessionPlannerDisplayConcurrentPreservesEverySession' -count=1
- git diff --check
